requirements.txt
alive-progress==1.6.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
coinapi.rest.v1==1.3
cycler==0.10.0
Flask==2.0.0
greenlet==1.0.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
jsonlib-python3==1.6.1
kaleido==0.2.1
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.0
matplotlib==3.4.2
mysql-connector-python==8.0.23
numpy==1.20.3
pandas==1.2.4
pandas-datareader==0.9.0
Pillow==8.2.0
plotly==4.14.3
progress==1.5
progressbar==2.5
progressbar2==3.53.1
protobuf==3.15.8
psutil==5.8.0
pyaes==1.6.1
pyasn1==0.4.8
pyodbc==4.0.30
pyparsing==2.4.7
pyTelegramBotAPI==3.7.7
python-coinmarketcap==0.2
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2021.1
requests==2.25.1
retrying==1.3.3
rsa==4.7.2
schedule==1.1.0
six==1.15.0
SQLAlchemy==1.4.9
telebot==0.0.4
Telethon==1.21.1
tqdm==4.60.0
urllib3==1.26.4
Werkzeug==2.0.0
